// src/app/ef-management/enum/ef-types.enum.ts
export enum efType {
    Common = 'common', 
    Fuel = "fuel",
    FuelFactor = "FuelFactor",
    FuelPrice ="FuelPrice",
    FuelSpecification = "FuelSpecification",
    FreightWater = "FreightWater",
    FreightRail = "FreightRail",
    MunicipalWaterTariff = "MunicipalWaterTariff",
    Defra = "Defra"

}